add test if files exists

+++ using-torch-on-hadoop-assets/countInput-run.sh
@@ -1,16 +1,23 @@
 # run countInput map-reduce streaming job
+# change to the directory that contains the source and shell file
+# then run
+#  ./countInput-run.sh
 
 # terminate after first line that fails
 set -e
 
+# print each command before executing it
+set -x
+
 # set control variables
 INPUT_FILE="parcels-HEATING.CODE-known-val.pairs"
+INPUT_FILE="courant-abel-prize-winners.txt"
 JOB_NAME="countInput"
 HPC_ID="rel292"
 
 # build other variables
 USER_DIR="/user/$USER"
-SRC="$HOME/re"
+SRC="$HOME/using-torch-on-hadoop-assets"
 SRC=$PWD
 MAPPER="${SRC}/${JOB_NAME}-map.lua"
 REDUCER="${SRC}/${JOB_NAME}-reduce.lua"
@@ -34,18 +41,38 @@ STREAMING="hadoop-streaming-1.0.3.16.jar"
 #which hadoop
 #STREAMING="hadoop-streaming-1.1.2.jar"
 
-# delete output from previous run
+# copy test file to the hadoop file system if it is not already there
+hadoop fs -test -e $INPUT_FILE
+echo rc from test of input file=$?
+if [ $? -ne 0 ]
+then
+  echo copying input from local file system to hadoop file system
+  hadoop fs -copyFromLocal $INPUT_FILE $INPUT_FILE
+else
+  echo input file already in the hadoop file system
+fi
+
+
+# delete output from previous run if it exists
 # NOTE: these command must be commented out when the script is first run, as
 # the delete command fails if the directory does not exist
 # and the mkdir fails if the directory already exists
-echo deleting $OUTPUT_DIR
-hadoop fs -rmr $OUTPUT_DIR
+echo about to test directory $OUTPUT_DIR
+hadoop fs -test -e $OUTPUT_DIR
+echo rc from test of output dir=$?
+if [ $? -eq 0 ]
+then
+  echo deleting output directory: $OUTPUT_DIR
+  hadoop fs -rmr $OUTPUT_DIR
+else
+  echo output directory not in the hadoop file system
+fi
 
 # create output directory
 #echo creating output directory $OUTPUT_DIR
 #hadoop fs -mkdir ${OUTPUT_DIR}
 
-# create output directory using streaming interface
+# run the streaming job; it will create the output directory
 echo
 echo
 echo creating output dirctory using streaming interface
